Important change: converted all short data fields to int and float to double in DATACH and DATABR structures; changed the class functions accordingly; old binary DBR and DCH files are no more readable (DK)
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSTANDALONE/trunk@251 9ff7d5ee-7a19-0410-ab54-ec97d5369f81

+++ GEMIPM2K/ms_param.cpp
@@ -223,10 +223,10 @@ void TMulti::CompG0Load()
    else
    {
        pmp->denW = LagranInterp( dCH->Pval, dCH->TCval, dCH->roW,
-                          (float)P, (float)TC, dCH->nTp, dCH->nPp,1 );
+                          P, TC, dCH->nTp, dCH->nPp,1 );
         //       pmp->denWg = tpp->RoV;
        pmp->epsW = LagranInterp( dCH->Pval, dCH->TCval, dCH->epsW,
-                          (float)P, (float)TC, dCH->nTp, dCH->nPp,1 );
+                          P, TC, dCH->nTp, dCH->nPp,1 );
        //       pmp->epsWg = tpp->EpsV;
    }
  }
@@ -259,9 +259,9 @@ void TMulti::CompG0Load()
      else
      {
        Go = LagranInterp( dCH->Pval, dCH->TCval, dCH->G0+jj,
-                          (float)P, (float)TC, dCH->nTp, dCH->nPp,1 );
+                          P, TC, dCH->nTp, dCH->nPp,1 );
        Vv = LagranInterp( dCH->Pval, dCH->TCval, dCH->V0+jj,
-                          (float)P, (float)TC, dCH->nTp, dCH->nPp, 1 );
+                          P, TC, dCH->nTp, dCH->nPp, 1 );
      }
      if( pmp->tpp_G )
     	  pmp->tpp_G[j] = Go;
